DTR/DSR signal inversion fix
Bug:

DS1100 configured with

a) Connection mode: 3 - On command OR DSR=HIGH
b) DTR mode:  1- Indicate connection status

DTR/DSR signalling is inverted in the ds4.38 binary as tested on the DS1100

Detail:

The DS203 connects to the configured host when its DSR input reads high, and it indicates successful connection by asserting its DTR output.
The DS203 disconnects to the configured host when its DSR input reads low, and it indicates it is no longer connected to the host by de-asserting its DTR output.

The DS1100 running ds4.38 is behaving exactly the opposite:
The DS1100 connects to the configured host when its DSR input reads low, and it indicates successful connection by de-asserting its DTR output.
The DS1100 disconnects to the configured host when its DSR input reads high, and it indicates it is no longer connected to the host by asserting its DTR output.

This behavior permeates to the floating signal state; when no device is attached to the DS1100 it thinks it should connect to its configured host.  This is the opposite of earlier DS203 behavior, and presents a compatibility problem with existing applications expecting the prior behavior of the DS203, requiring both that a device be connected with the SOI device and requesting a host connection (by asserting DTR) before the the DS1100 attempts to connect with the remote internet host.

This update restores the classic, expected behavior.

New Binary compiled as ds4.39